infras_core/utils/io.py
"""I/O utilities for IMS and MiniSEED data"""

import numpy as np
from typing import Optional, Tuple, List
import json
import yaml


class IMSDataLoader:
    """Loader for IMS infrasound data"""
    
    def __init__(self, data_dir: str):
        self.data_dir = data_dir
    
    def load_station(self, station_id: str, start_time: str, end_time: str) -> np.ndarray:
        """
        Load data for a specific station
        
        Parameters
        ----------
        station_id : str
            IMS station ID (e.g., 'IS42')
        start_time : str
            Start time in ISO format
        end_time : str
            End time in ISO format
        
        Returns
        -------
        np.ndarray
            Pressure time series
        """
        # TODO: Implement actual data loading
        return np.random.randn(3600 * 20)  # 1 hour at 20 Hz
    
    def load_array(self, station_ids: List[str], **kwargs) -> dict:
        """Load data for multiple stations"""
        data = {}
        for sid in station_ids:
            data[sid] = self.load_station(sid, **kwargs)
        return data


def read_miniseed(filename: str) -> Tuple[np.ndarray, float]:
    """
    Read MiniSEED file
    
    Parameters
    ----------
    filename : str
        Path to MiniSEED file
    
    Returns
    -------
    data : np.ndarray
        Time series data
    fs : float
        Sampling frequency (Hz)
    """
    # TODO: Implement with ObsPy
    return np.random.randn(3600), 20.0


def load_station_metadata(filename: str) -> dict:
    """Load station metadata from YAML/JSON"""
    if filename.endswith('.json'):
        with open(filename, 'r') as f:
            return json.load(f)
    elif filename.endswith(('.yaml', '.yml')):
        with open(filename, 'r') as f:
            return yaml.safe_load(f)
    else:
        raise ValueError(f"Unsupported file format: {filename}")
